Add setting to show or hide export buttons
My intended use for this plugin is mainly just to view .note files, not export them, so I would prefer not to have these buttons take up space in every file. I think a better implementation might be to have these functions available on hover or right-click, but this works for now to get that screen real estate back.

Default behavior is to show the buttons, same as before this commit.