Add setting to collapse recognized text when viewing .note files
This applies to viewing .note files only, and not to generated markdown.

When viewing notes, the recognized text is sometimes useful, but oftentimes it has many errors, so I'd prefer to hide it and focus on the image of the note instead. Putting it underneath an HTML `details` element tucks it away while keeping it available if needed.