Add container for each page. Setting now limits both width and height instead of just width. Multiple pages can be visible horizontally.
I think this should play nice with horizontal pages, but am unable to test because horizontal pages aren't working quite yet.

Also changed the default setting to 800px, which is a nice medium size to be legible but not too big. Max setting now goes to 1900 px (slightly larger than the 1872 px long edge of a Nomad page).