fix: Add margin around page container (#38)
Add margin around page container so that the recognized text of one page doesn't run into that of the next page. 1.2em seems like a reasonable amount for readability. Could maybe go slightly higher.

I also added some margin on the bottom of the recognized text so it's not too close to the image of the page.